GCC/Clang Fixes
General:
* replaced usage of integral literal suffixes with standard c++ suffixes
* reworked alias type-usings with corresponding platform type
* fixed forceinline macro for gcc and clang

+++ src/Editor.UI/Helper/AssetBrowserHelper.cpp
@@ -204,7 +204,7 @@ ptr<engine::assets::TextureAsset> AssetBrowserHelper::getItemIconForDirectory(cr
     for (const auto& entry : _directoryIcons) {
         if (name.contains(entry.first) && entry.first.size() > length) {
 
-            length = entry.first.size();
+            length = static_cast<s32>(entry.first.size());
             tex = entry.second;
 
             if (length == name.size()) {
